refactor(cli): move reintegrate-buffer handler to its own file
Extracts the ReintegrateBuffer handler body into reintegrate_buffer.rs as
a standalone function, keeping cli.rs's match arm a one-line call. Gives
the follow-up progress-bar PR a self-contained file to build on.
